Add `forge_vertex` mode allowing for extremely quick diagram layout
`forge_vertex`, switched on by default, allows you to draw an edge between empty cells (technically *to* an empty cell, but a new behaviour allows to drag from an empty one too), which will create the necessary vertices and edges. This allows diagrams to be created simply by dragging (and concrentrating primarily on the edges rather than the vertices).