Update the workflow for pushing to GitHub Pages
There is now a three-stage process. `master` is always the latest version of the code, but isn't live anywhere. `dev` is live under `varkor.github.io/quiver/dev` and will be used for testing. `release` is live under `varkor.github.io/quiver` and is the current version of quiver. We move code `master` → `dev` → `release`.