Replace native sliders with a custom implementation
This isn't necessary for the current sliders, though it does have the advantage of making their behaviour consistent across browsers, and fixes an issue where slider values are not visible in Firefox. However, more importantly, it opens up the possibility of sliders with multiple thumbs, which are necessary for asymmetric edge shortening.